requirements.txt
argcomplete==3.0.8
beautifultable==1.1.0
dill==0.3.6
h5py==3.7.0
matplotlib==3.7.0
numpy==1.23.5
Pygments==2.11.2
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.2.1
scipy==1.10.0
simple_term_menu==1.6.1
termcolor==2.3.0
torch==2.0.0
tqdm==4.64.1
